Cleaned up routes files slightly
